The query mod selector can now show the selected entity and the label in either order (ex: Sort By A-Z: "Sort By" as the label and "A-Z" as the selection and 20 Results Per Page: "20" as the selection and "Results Per Page" as the label.
Game Cards without images no longer break the site.

useQueryController cuts off the last ampersand before the key.

Query Mod Selector can now also send the selected value derived from the selected item's key to the useQueryController, but then use the text from the contents of the selected item to change the label on the selector itself.

To reiterate, the goal here is to have QueryModSelector be a single component that will handle all dropdowns or inputs that update the game query, with the exception of genres, because genres is using too different of a design.